+++ torchao/quantization/pt2e/pt2e/duplicate_dq_pass.py
@@ -12,9 +12,11 @@ import torch
 from torch.fx.node import map_arg
 from torch.fx.passes.infra.pass_base import PassBase, PassResult
 
-from torchao.quantization.pt2e.pt2e.utils import (
-    _filter_sym_size_users,
+from torchao.quantization.pt2e.quantizer.utils import (
     _is_valid_annotation,
+)
+from torchao.quantization.pt2e.utils import (
+    _filter_sym_size_users,
 )
 
 logger = logging.getLogger(__name__)
