+++ torchao/quantization/pt2e/pt2e/port_metadata_pass.py
@@ -12,12 +12,14 @@ import torch
 from torch._export.error import InternalError
 from torch.fx.passes.infra.pass_base import PassBase, PassResult
 
-from torchao.quantization.pt2e.pt2e.utils import (
-    _filter_sym_size_users,
-    _find_q_dq_node_for_user,
+from torchao.quantization.pt2e.quantizer import QuantizationSpecBase
+from torchao.quantization.pt2e.quantizer.utils import (
     _is_valid_annotation,
 )
-from torchao.quantization.pt2e.quantizer import QuantizationSpecBase
+from torchao.quantization.pt2e.utils import (
+    _filter_sym_size_users,
+    _find_q_dq_node_for_user,
+)
 from torchao.quantization.quant_primitives import quant_lib  # noqa: F401
 from torchao.utils import TORCH_VERSION_AT_LEAST_2_5
